(chores) camel-kafka: rework the commit management
Breaks the commit management for noop, sync and async strategies and
streamline the equivalent manual commit classes so that responsibilities
boundaries are clearer
