+++ components/camel-kafka/src/main/java/org/apache/camel/component/kafka/consumer/NoopCommitManager.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.camel.component.kafka.consumer;
+
+import org.apache.camel.component.kafka.KafkaConsumer;
+import org.apache.kafka.clients.consumer.Consumer;
+import org.apache.kafka.common.TopicPartition;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class NoopCommitManager extends AbstractCommitManager {
+    private static final Logger LOG = LoggerFactory.getLogger(NoopCommitManager.class);
+
+    public NoopCommitManager(Consumer<?, ?> consumer, KafkaConsumer kafkaConsumer, String threadId, String printableTopic) {
+        super(consumer, kafkaConsumer, threadId, printableTopic);
+    }
+
+    @Override
+    public void commit() {
+        LOG.info("Auto commit on stop {} from {} is disabled (none)", threadId, printableTopic);
+
+    }
+
+    @Override
+    public void commitOffsetOnStop(TopicPartition partition, long partitionLastOffset) {
+        if (LOG.isDebugEnabled()) {
+            LOG.debug("Auto commit on stop {} from topic {} is disabled (none)", threadId, partition.topic());
+        }
+    }
+
+    @Override
+    public void commitOffset(TopicPartition partition, long partitionLastOffset) {
+        if (LOG.isDebugEnabled()) {
+            LOG.debug("Auto commit on stop {} from topic {} is disabled (none)", threadId, partition.topic());
+        }
+    }
+}
